clean up new Umfpack caching
