+++ SRC/system_of_eqn/linearSOE/umfGEN/UmfpackSolver02.cpp
@@ -1,7 +1,7 @@
 
 
 #include <UmfpackSolver02.h>
-#include <UmfpackLinSOE02.h>
+#include <SparseGenCSC.h>
 
 #include <Channel.h>
 #include <Constants.h>
@@ -9,132 +9,125 @@
 #include <Vector.h>
 
 UmfpackSolver02::UmfpackSolver02(bool doDet)
-  : LinearSOESolver(SOLVER_TAGS_UmfpackSolver02), symbolic(nullptr),
+  : LinearSOESolver(-1), symbolic(nullptr),
     control(), info(), theSOE(nullptr), determinant(0.0), doDeterminant(doDet)
 {
 }
 
 UmfpackSolver02::~UmfpackSolver02()
 {
-    this->clearSymbolic();
+  this->clearSymbolic();
 }
 
 int
 UmfpackSolver02::solve()
 {
-    if (theSOE == nullptr) {
-        return -1;
-    }
+  if (theSOE == nullptr)
+    return -1;
 
-    return this->solve(theSOE->B, theSOE->X);
+  return this->solve(theSOE->getB(), theSOE->X);
 }
 
 int
 UmfpackSolver02::solve(const Vector &B, Vector &X)
 {
-    if (theSOE == nullptr) {
-        return -1;
-    }
+  if (theSOE == nullptr)
+    return -1;
 
-    const int size = X.Size();
-    const int nnz = static_cast<int>(theSOE->Ai.size());
-    if (size == 0 || nnz == 0) {
-        return 0;
-    }
+  const int size = X.Size();
+  const int nnz = static_cast<int>(theSOE->Ai.size());
+  if (size == 0 || nnz == 0)
+    return 0;
 
-    if (symbolic == nullptr) {
-        return -1;
-    }
+  if (symbolic == nullptr)
+    return -1;
 
-    if (B.Size() != size || static_cast<int>(theSOE->Ap.size()) != size + 1) {
-        return -1;
-    }
+  if (B.Size() != size || static_cast<int>(theSOE->Ap.size()) != size + 1)
+    return -1;
 
-    void *numeric = nullptr;
-    int status = umfpack_di_numeric(theSOE->Ap.data(), theSOE->Ai.data(),
-                                    theSOE->Ax.data(), symbolic, &numeric,
-                                    control.data(), info.data());
-    if (status != UMFPACK_OK) {
-        if (numeric != nullptr) {
-            umfpack_di_free_numeric(&numeric);
-        }
-        return -1;
-    }
+  void *numeric = nullptr;
+  int status = umfpack_di_numeric(theSOE->Ap.data(), theSOE->Ai.data(),
+                                  theSOE->Ax.data(), symbolic, &numeric,
+                                  control.data(), info.data());
+  if (status != UMFPACK_OK) {
+    if (numeric != nullptr)
+      umfpack_di_free_numeric(&numeric);
+    return -1;
+  }
 
-    status = umfpack_di_solve(UMFPACK_A, theSOE->Ap.data(), theSOE->Ai.data(),
-                              theSOE->Ax.data(), &X(0), &B(0), numeric,
-                              control.data(), info.data());
+  status = umfpack_di_solve(UMFPACK_A, theSOE->Ap.data(), theSOE->Ai.data(),
+                            theSOE->Ax.data(), &X(0), &B(0), numeric,
+                            control.data(), info.data());
 
-    if (doDeterminant) {
-        umfpack_di_get_determinant(&determinant, nullptr, numeric, info.data());
-    }
+  if (doDeterminant)
+    umfpack_di_get_determinant(&determinant, nullptr, numeric, info.data());
 
-    umfpack_di_free_numeric(&numeric);
-    return status == UMFPACK_OK ? 0 : -1;
+  umfpack_di_free_numeric(&numeric);
+  return status == UMFPACK_OK ? 0 : -1;
 }
 
 int
 UmfpackSolver02::setSize()
 {
-    this->clearSymbolic();
-    umfpack_di_defaults(control.data());
-    control[UMFPACK_PIVOT_TOLERANCE] = 1.0;
-    control[UMFPACK_STRATEGY] = UMFPACK_STRATEGY_SYMMETRIC;
+  this->clearSymbolic();
+  umfpack_di_defaults(control.data());
+  control[UMFPACK_PIVOT_TOLERANCE] = 1.0;
+  control[UMFPACK_STRATEGY] = UMFPACK_STRATEGY_SYMMETRIC;
 
-    if (theSOE == nullptr) {
-        return -1;
-    }
+  if (theSOE == nullptr) {
+    return -1;
+  }
 
-    const int size = theSOE->X.Size();
-    const int nnz = static_cast<int>(theSOE->Ai.size());
-    if (size == 0 || nnz == 0) {
-        return 0;
-    }
-
-    if (static_cast<int>(theSOE->Ap.size()) != size + 1) {
-        return -1;
-    }
-
-    const int status = umfpack_di_symbolic(size, size, theSOE->Ap.data(),
-                                           theSOE->Ai.data(), theSOE->Ax.data(),
-                                           &symbolic, control.data(), info.data());
-    if (status != UMFPACK_OK) {
-        this->clearSymbolic();
-        return -1;
-    }
-
+  const int size = theSOE->X.Size();
+  const int nnz = static_cast<int>(theSOE->Ai.size());
+  if (size == 0 || nnz == 0)
     return 0;
+
+  if (static_cast<int>(theSOE->Ap.size()) != size + 1)
+    return -1;
+
+  const int status = umfpack_di_symbolic(size, size,
+                                         theSOE->Ap.data(),
+                                         theSOE->Ai.data(),
+                                         theSOE->Ax.data(),
+                                         &symbolic, control.data(), info.data());
+  if (status != UMFPACK_OK) {
+    this->clearSymbolic();
+    return -1;
+  }
+
+  return 0;
 }
 
 int
-UmfpackSolver02::setLinearSOE(UmfpackLinSOE02 &soe)
+UmfpackSolver02::setLinearSOE(SparseGenCSC &soe)
 {
-    theSOE = &soe;
-    return 0;
+  theSOE = &soe;
+  return 0;
 }
 
 int
 UmfpackSolver02::sendSelf(int, Channel &)
 {
-    return 0;
+  return 0;
 }
 
 int
 UmfpackSolver02::recvSelf(int, Channel &, FEM_ObjectBroker &)
 {
-    return 0;
+  return 0;
 }
 
 double
 UmfpackSolver02::getDeterminant()
 {
-    return doDeterminant ? determinant : OpenSees::Constants::nan;
+  return doDeterminant ? determinant : OpenSees::Constants::nan;
 }
 
 void
 UmfpackSolver02::clearSymbolic()
 {
-    if (symbolic != nullptr) {
-        umfpack_di_free_symbolic(&symbolic);
-    }
+  if (symbolic != nullptr) {
+    umfpack_di_free_symbolic(&symbolic);
+  }
 }
